requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
Django==4.2
django-channels==0.7.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
msgpack==1.0.8
oauthlib==3.2.2
PyJWT==2.8.0
pytz==2024.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
